+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add a CSS class to move content offscreen so it's only available for users
+  of assistive technologies.
+
 ## [1.6.1] - 2019-08-03
 
 ### Fixed